+++ evalml/automl/automl_search.py
@@ -111,6 +111,7 @@ def search(
     patience=None,
     tolerance=None,
     problem_configuration=None,
+    n_splits=3,
     verbose=False,
 ):
     """Given data and configuration, run an automl search.
@@ -137,7 +138,8 @@ def search(
         tolerance (float): Minimum percentage difference to qualify as score improvement for early stopping.
             Only applicable if patience is not None. Defaults to None.
         problem_configuration (dict): Additional parameters needed to configure the search. For example,
-            in time series problems, values should be passed in for the date_index, gap, and max_delay variables.
+            in time series problems, values should be passed in for the time_index, gap, forecast_horizon, and max_delay variables.
+        n_splits (int): Number of splits to use with the default data splitter.
         verbose (boolean): Whether or not to display semi-real-time updates to stdout while search is running. Defaults to False.
 
     Returns:
@@ -150,20 +152,10 @@ def search(
     y_train = infer_feature_types(y_train)
     problem_type = handle_problem_types(problem_type)
 
-    datetime_column = None
     if is_time_series(problem_type):
-        if problem_configuration:
-            if "date_index" in problem_configuration:
-                datetime_column = problem_configuration["date_index"]
-            else:
-                raise ValueError(
-                    "For the default data checks to run in time series, date_index has to be passed in problem_configuration. "
-                    f"Received {problem_configuration}"
-                )
-        else:
-            raise ValueError(
-                "For the default data checks to run in time series, the problem_configuration parameter must be specified."
-            )
+        is_valid, msg = contains_all_ts_parameters(problem_configuration)
+        if not is_valid:
+            raise ValueError(msg)
 
     if objective == "auto":
         objective = get_default_primary_search_objective(problem_type)
@@ -180,6 +172,14 @@ def search(
     elif mode == "long" and max_time is None:
         max_batches = 6  # corresponds to end of 'long' exploration phase
 
+    data_splitter = make_data_splitter(
+        X=X_train,
+        y=y_train,
+        problem_type=problem_type,
+        problem_configuration=problem_configuration,
+        n_splits=n_splits,
+    )
+
     automl_config = {
         "X_train": X_train,
         "y_train": y_train,
@@ -190,10 +190,15 @@ def search(
         "patience": patience,
         "tolerance": tolerance,
         "verbose": verbose,
+        "problem_configuration": problem_configuration,
+        "data_splitter": data_splitter,
     }
 
     data_checks = DefaultDataChecks(
-        problem_type=problem_type, objective=objective, datetime_column=datetime_column
+        problem_type=problem_type,
+        objective=objective,
+        n_splits=n_splits,
+        problem_configuration=problem_configuration,
     )
     data_check_results = data_checks.validate(X_train, y=y_train)
     if len(data_check_results.get("errors", [])):
@@ -210,6 +215,7 @@ def search_iterative(
     problem_type=None,
     objective="auto",
     problem_configuration=None,
+    n_splits=3,
     **kwargs,
 ):
     """Given data and configuration, run an automl search.
@@ -227,7 +233,8 @@ def search_iterative(
             - LogLossMulticlass for multiclass classification problems, and
             - R2 for regression problems.
         problem_configuration (dict): Additional parameters needed to configure the search. For example,
-            in time series problems, values should be passed in for the date_index, gap, forecast_horizon, and max_delay variables.
+            in time series problems, values should be passed in for the time_index, gap, forecast_horizon, and max_delay variables.
+        n_splits (int): Number of splits to use with the default data splitter.
         **kwargs: Other keyword arguments which are provided will be passed to AutoMLSearch.
 
     Returns:
@@ -240,24 +247,22 @@ def search_iterative(
     y_train = infer_feature_types(y_train)
     problem_type = handle_problem_types(problem_type)
 
-    datetime_column = None
     if is_time_series(problem_type):
-        if problem_configuration:
-            if "date_index" in problem_configuration:
-                datetime_column = problem_configuration["date_index"]
-            else:
-                raise ValueError(
-                    "For the default data checks to run in time series, date_index has to be passed in problem_configuration. "
-                    f"Received {problem_configuration}"
-                )
-        else:
-            raise ValueError(
-                "For the default data checks to run in time series, the problem_configuration parameter must be specified."
-            )
+        is_valid, msg = contains_all_ts_parameters(problem_configuration)
+        if not is_valid:
+            raise ValueError(msg)
 
     if objective == "auto":
         objective = get_default_primary_search_objective(problem_type)
     objective = get_objective(objective, return_instance=False)
+
+    data_splitter = make_data_splitter(
+        X=X_train,
+        y=y_train,
+        problem_type=problem_type,
+        problem_configuration=problem_configuration,
+        n_splits=n_splits,
+    )
 
     automl_config = kwargs
     automl_config.update(
@@ -267,11 +272,16 @@ def search_iterative(
             "problem_type": problem_type,
             "objective": objective,
             "max_batches": 1,
+            "problem_configuration": problem_configuration,
+            "data_splitter": data_splitter,
         }
     )
 
     data_checks = DefaultDataChecks(
-        problem_type=problem_type, objective=objective, datetime_column=datetime_column
+        problem_type=problem_type,
+        objective=objective,
+        n_splits=n_splits,
+        problem_configuration=problem_configuration,
     )
     data_check_results = data_checks.validate(X_train, y=y_train)
     if len(data_check_results.get("errors", [])):
@@ -360,7 +370,7 @@ class AutoMLSearch:
             max_iterations have precedence over stopping the search.
 
         problem_configuration (dict, None): Additional parameters needed to configure the search. For example,
-            in time series problems, values should be passed in for the date_index, gap, forecast_horizon, and max_delay variables.
+            in time series problems, values should be passed in for the time_index, gap, forecast_horizon, and max_delay variables.
 
         train_best_pipeline (boolean): Whether or not to train the best pipeline before returning it. Defaults to True.
 
@@ -605,8 +615,8 @@ class AutoMLSearch:
         self._best_pipeline = None
         self._searched = False
 
-        self.X_train = infer_feature_types(X_train)
-        self.y_train = infer_feature_types(y_train)
+        self.X_train = infer_feature_types(X_train, ignore_nullable_types=True)
+        self.y_train = infer_feature_types(y_train, ignore_nullable_types=True)
 
         default_data_splitter = make_data_splitter(
             self.X_train,
@@ -795,8 +805,6 @@ class AutoMLSearch:
             is_valid, msg = contains_all_ts_parameters(problem_configuration)
             if not is_valid:
                 raise ValueError(msg)
-            if problem_configuration["date_index"] is None:
-                raise ValueError("date_index cannot be None!")
         return problem_configuration or {}
 
     def _handle_keyboard_interrupt(self):
@@ -1091,9 +1099,9 @@ class AutoMLSearch:
         else:
             gap = self.problem_configuration["gap"]
             forecast_horizon = self.problem_configuration["forecast_horizon"]
-            date_index = self.problem_configuration["date_index"]
+            time_index = self.problem_configuration["time_index"]
             baseline = make_timeseries_baseline_pipeline(
-                self.problem_type, gap, forecast_horizon, date_index
+                self.problem_type, gap, forecast_horizon, time_index
             )
         return baseline