+++ evalml/data_checks/__init__.py
@@ -23,3 +23,4 @@ from .natural_language_nan_data_check import NaturalLanguageNaNDataCheck
 from .target_distribution_data_check import TargetDistributionDataCheck
 from .datetime_format_data_check import DateTimeFormatDataCheck
 from .ts_parameters_data_check import TimeSeriesParametersDataCheck
+from .ts_splitting_data_check import TimeSeriesSplittingDataCheck