+++ evalml/pipelines/regression_pipeline.py
@@ -26,7 +26,10 @@ class RegressionPipeline(PipelineBase):
         ...
         >>> assert pipeline.custom_name == "My Regression Pipeline"
         >>> assert pipeline.component_graph.component_dict.keys() == {'Simple Imputer', 'Linear Regressor'}
-        ...
+
+        The pipeline parameters will be chosen from the default parameters for every component, unless specific parameters
+        were passed in as they were above.
+
         >>> assert pipeline.parameters == {
         ...     'Simple Imputer': {'impute_strategy': 'most_frequent', 'fill_value': None},
         ...     'Linear Regressor': {'fit_intercept': True, 'normalize': True, 'n_jobs': -1}}