+++ docs/source/release_notes.rst
@@ -2,6 +2,28 @@ Release Notes
 -------------
 **Future Releases**
     * Enhancements
+        * Added ``TimeSeriesSplittingDataCheck`` to ``DefaultDataChecks`` to verify adequate class representation in time series classification problems :pr:`3141`
+        * Added the ability to accept serialized features and skip computation in ``DFSTransformer`` :pr:`3106`
+        * Added support for known-in-advance features :pr:`3149`
+    * Fixes
+        * Fixed error caused when tuning threshold for time series binary classification :pr:`3140`
+    * Changes
+        * ``TimeSeriesParametersDataCheck`` was added to ``DefaultDataChecks`` for time series problems :pr:`3139`
+        * Renamed ``date_index`` to ``time_index`` in ``problem_configuration`` for time series problems :pr:`3137`
+    * Documentation Changes
+        * Added comments to provide clarity on doctests :pr:`3155`
+    * Testing Changes
+        * Parameterized tests in ``test_datasets.py`` :pr:`3145`
+
+
+.. warning::
+
+    **Breaking Changes**
+        * Renamed ``date_index`` to ``time_index`` in ``problem_configuration`` for time series problems :pr:`3137`
+
+
+**v0.39.0 Dec. 9, 2021**
+    * Enhancements
         * Renamed ``DelayedFeatureTransformer`` to ``TimeSeriesFeaturizer`` and enhanced it to compute rolling features :pr:`3028`
         * Added ability to impute only specific columns in ``PerColumnImputer`` :pr:`3123`
         * Added ``TimeSeriesParametersDataCheck`` to verify the time series parameters are valid given the number of splits in cross validation :pr:`3111`
@@ -13,6 +35,7 @@ Release Notes
         * Cleaned up and refactored ``InvalidTargetDataCheck`` implementation and docstring :pr:`3122`
         * Removed indices information from the output of ``HighlyNullDataCheck``'s ``validate()`` method :pr:`3092`
         * Added ``ReplaceNullableTypes`` component to prepare for handling pandas nullable types. :pr:`3090`
+        * Updated ``make_pipeline`` for handling pandas nullable types in preprocessing pipeline. :pr:`3129`
         * Removed unused ``EnsembleMissingPipelinesError`` exception definition :pr:`3131`
     * Documentation Changes
     * Testing Changes